refactor(101645): :recycle: separa o codigo do relatorio em excel
